+++ CloudShell.ResourceDefinitions.ResourceManager/ResourceModelResourceManagerEndpointProjectionProvider.cs
@@ -1,0 +1,9 @@
+using ResourceModelResource = CloudShell.ResourceDefinitions.Resource;
+
+namespace CloudShell.ResourceDefinitions.ResourceManager;
+
+public interface IResourceModelResourceManagerEndpointProjectionProvider
+{
+    ResourceModelResourceManagerEndpointProjection? GetEndpointProjection(
+        ResourceModelResource resource);
+}